pyproject.toml
[tool.poetry]
name = "unlock-auto"
version = "0.1.0"
description = "Admin"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
bcrypt = "^4.0.1"
django-behaviors = "^0.5.1"
django-environ = "^0.11.2"
django-ipware = "^5.0.0"
django-axes = "^6.1.0"
whitenoise = "^6.5.0"
django-storages = "^1.13.2"
djangorestframework = "^3.14.0"
djangorestframework-camel-case = "^1.4.2"
drf-spectacular = "^0.26.4"
django-filter = "^23.2"
django-split-settings = "^1.2.0"
django-healthchecks = "^1.5.0"
redis = "^5.0.0"
sentry-sdk = "^1.29.2"
pillow = "^10.0.0"
gunicorn = "^21.2.0"
gevent = "^23.7.0"
celery = "^5.3.1"
psycopg2-binary = "^2.9.7"
boto3 = "^1.24.80"
djangorestframework-simplejwt = "5.3.0"
django-ckeditor = "^6.7.0"
uuid = "^1.30"
django-nested-admin = "^4.0.2"
django-fsm = "^2.8.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.2.0"}
pytest-django = "^4.6.0"
django-import-export = "^3.3.3"
aiohttp = "^3.9.1"
django-autocomplete-light = "^3.9.7"
django-request-logging = "^0.7.5"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.5.2"
pytest-deadfixtures = "^2.2.1"
pytest-env = "^0.6.2"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.8.2"
freezegun = "^1.2.2"
jedi = "^0.18.1"
flake8-absolute-import = "^1.0.0.1"
flake8-bugbear = "^22.9.23"
flake8-cognitive-complexity = "^0.1.0"
flake8-commas = "^2.1.0"
flake8-django = "^1.1.5"
flake8-eradicate = "^1.4.0"
flake8-isort = "^4.2.0"
flake8-fixme = "^1.1.1"
flake8-mock = "^0.3"
pytest-randomly = "^3.12.0"
mixer = "^7.2.2"
flake8-multiline-containers = "^0.0.19"
flake8-mutable = "^1.2.0"
flake8-pep3101 = "^1.3.0"
flake8-pie = "^0.16.0"
flake8-print = "^5.0.0"
flake8-printf-formatting = "^1.1.2"
flake8-pytest = "^1.4"
flake8-pytest-style = "^1.6.0"
flake8-quotes = "^3.3.1"
flake8-simplify = "^0.19.3"
flake8-todo = "^0.7"
flake8-use-fstring = "^1.4"
flake8-walrus = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
